requirements.txt
amqp==2.5.0
billiard==3.6.0.0
celery==4.3.0
Django==2.2.3
django-celery-beat==1.5.0
django-celery-results==1.1.2
django-filter==2.1.0
django-safedelete==0.5.1
django-timezone-field==3.0
djangorestframework==3.10.0
djangorestframework-simplejwt==4.3.0
kombu==4.6.3
Markdown==3.1.1
PyJWT==1.7.1
python-crontab==2.3.8
python-dateutil==2.8.0
pytz==2019.1
redis==3.2.1
six==1.12.0
sqlparse==0.3.0
vine==1.3.0
